Fixed "unable to reach node" issue
Made the pickup/dropoff functions get called for a longer period of time to ensure all passengers get dropped off.

The pickup function now checks to see if the passenger destination is reachable from where it is

i and j get set to 0 at the end of each function call (this might be unnecessary)

removed the check for shortest path length from the dropoff function. This was causing the 'unable to reach node' issue and is no longer required since this is now being checked in the pickup function

Passenger is now dropped off by the correct driver

Only drops off passengers that have been picked up